// src/app/models/home/prices-details.model.ts
export interface PricesDetails {
    pricesKiloItens: string[];
    pricesKiloValues: string[];

    pricesFemaleItens: string[];
    pricesFemaleValues: string[];

    pricesMaleItens: string[];
    pricesMaleValues: string[];

    pricesChildrenItens: string[];
    pricesChildrenValues: string[];

    pricesCouchItens: string[];
    pricesCouchValues: string[];

    pricesRugItens: string[];
    pricesRugValues: string[];

    pricesCurtainItens: string[];
    pricesCurtainValues: string[];

    pricesBedItens: string[];
    pricesBedValues: string[];

    pricesTrussardiItens: string[];
    pricesTrussardiValues: string[];
    
    pricesOthersItens: string[];
    pricesOthersValues: string[];
}